requirements.txt
flask==3.1.0
PyJWT==2.10.1
werkzeug==3.1.3
tensorflow==2.18.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2
matplotlib==3.10.0
flask-cors==5.0.1
